+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add check to Prefix for ByRef not being array type
 * Fix ByRef not working on array types
 * Fix some weirdness of UI warnings
+* Add toggle all mods button
 
 #### v3.10.1 ####
 * Fix crash in incomplete last-update code